function nested loop task correction

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -11,7 +11,7 @@ int main(void)
 	long int immediate_no = 1;
 	long int previous_no = 0;
 	long int sum_of_immediate_and_previous_no = 0;
-	int counter;
+	long int counter;
 
 	for (counter = 0; counter < 50; counter++)
 	{
